*Partial update rAthena 507f047cfb. 1290826be2, 9c438ce762, 4527b66a5d

+++ trunk/src/map/clif.h
@@ -521,7 +521,7 @@ void clif_inventorylist(struct map_session_data *sd);
 void clif_equiplist(struct map_session_data *sd);
 
 void clif_cart_additem(struct map_session_data *sd,int n,int amount,int fail);
-void clif_cart_additem_ack(struct map_session_data *sd, int flag);
+void clif_cart_additem_ack(struct map_session_data *sd, uint8 flag);
 void clif_cart_delitem(struct map_session_data *sd,int n,int amount);
 void clif_cartlist(struct map_session_data *sd);
 void clif_clearcart(int fd);
